refactor: extract DAY constant and move MINUTE to shared space
